// src/service/user/article-draft/define/draft-list.d.ts
// Generated by https://quicktype.io

export interface DraftListJSON {
  artlist: Artlist | null
  code: number
  message: string
}

export interface Artlist {
  drafts: Draft[]
  page: Page
  draft_url: string
}

export interface Draft {
  id: number
  title: string
  content: string
  summary: string
  banner_url: string
  reason: string
  template_id: number
  state: number
  reprint: number
  image_urls: any[]
  origin_image_urls: any[]
  tags: any[]
  category: Category
  author: Author
  stats: null
  publish_time: number
  ctime: number
  mtime: number
  view_url: string
  edit_url: string
  is_preview: number
  dynamic_intro: string
}

export interface Author {
  mid: number
  name: string
  face: string
  pendant: Pendant
  official_verify: OfficialVerify
  nameplate: Nameplate
  vip: Vip
}

export interface Nameplate {
  nid: number
  name: string
  image: string
  image_small: string
  level: string
  condition: string
}

export interface OfficialVerify {
  type: number
  desc: string
}

export interface Pendant {
  pid: number
  name: string
  image: string
  expire: number
}

export interface Vip {
  type: number
  status: number
  due_date: number
  vip_pay_type: number
  theme_type: number
  label: null
  avatar_subscript: number
  nickname_color: string
}

export interface Category {
  id: number
  parent_id: number
  name: string
}

export interface Page {
  pn: number
  ps: number
  total: number
}